+++ src/main/java/kr/hhplus/be/ecommerce/interfaces/product/api/ProductController.java
@@ -1,7 +1,7 @@
-package kr.hhplus.be.ecommerce.interfaces.product;
+package kr.hhplus.be.ecommerce.interfaces.product.api;
 
-import kr.hhplus.be.ecommerce.application.product.ProductFacade;
-import kr.hhplus.be.ecommerce.application.product.ProductResult;
+import kr.hhplus.be.ecommerce.domain.product.ProductInfo;
+import kr.hhplus.be.ecommerce.domain.product.ProductService;
 import kr.hhplus.be.ecommerce.interfaces.ApiResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -11,11 +11,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class ProductController {
 
-    private final ProductFacade productFacade;
+    private final ProductService productService;
 
     @GetMapping("/api/v1/products")
     public ApiResponse<ProductResponse.Products> getProducts() {
-        ProductResult.Products products = productFacade.getProducts();
+        ProductInfo.Products products = productService.getSellingProducts();
         return ApiResponse.success(ProductResponse.Products.of(products));
     }
 }
